requirements.txt
streamlit
sqlalchemy
psycopg2-binary
pandas
opencv-python
face-recognition
numpy
pillow
torch
torchvision
scikit-learn
matplotlib
plotly
seaborn
streamlit-authenticator
bcrypt
python-multipart
email-validator
schedule
python-dotenv
qrcode[pil]
reportlab
openpyxl
xlsxwriter
streamlit-option-menu
streamlit-extras
streamlit-aggrid
streamlit-lottie
requests
psutil
cryptography
werkzeug 